requirements.txt
fastapi~=0.115.4
uvicorn==0.32.0
pydantic==2.9.2
requests==2.32.3
fuzzywuzzy==0.18.0
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2